+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.6.0] - 2025-08-27
+
 ### Added
 
 - cli: Add delay command to update todo due with datetime, time, or tomorrow keyword
@@ -17,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - cli: Update dashboard style
+
+## [v0.5.1] - 2025-08-19
+
+### Changed
+
+- cli: Show all events in today by default
+- core: `EventConditions.startable` expect a `DateTimeAnchor` now
+- core: `TodoConditions.due` expect a `DateTimeAnchor` now
 
 ### Fixed
 
@@ -210,7 +220,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add events and todos command
 - Add dashboard command
 
-[unreleased]: https://github.com/yzx9/aim/compare/v0.5.1...HEAD
+[unreleased]: https://github.com/yzx9/aim/compare/v0.6.0...HEAD
+[v0.6.0]: https://github.com/yzx9/aim/compare/v0.5.1...v0.6.0
 [v0.5.1]: https://github.com/yzx9/aim/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/yzx9/aim/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/yzx9/aim/compare/v0.3.2...v0.4.0